Analyste Fiscal/requirements.txt
crewai>=0.28.0
xero-python>=1.35.0
stripe>=7.0.0
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
cryptography>=41.0.0
schedule>=1.2.0
plotly>=5.17.0
dash>=2.14.0
python-docx>=1.1.0
PyPDF2>=3.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0
pydantic>=2.5.0
jinja2>=3.1.0
python-dateutil>=2.8.0
pytz>=2023.3
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
tabulate>=0.9.0
rich>=13.7.0
typer>=0.9.0
flask>=2.3.0
flask-uploads>=0.2.1 